implementing new UI with modal for creating bookings
