+++ app/components/booking/actions-dropdown.tsx
@@ -9,7 +9,7 @@ import {
 } from "~/components/shared/dropdown";
 import { useBookingStatus } from "~/hooks/use-booking-status";
 import { useUserIsSelfService } from "~/hooks/user-user-is-self-service";
-import type { BookingWithCustodians } from "~/routes/_layout+/bookings._index";
+import type { BookingWithCustodians } from "~/routes/_layout+/bookings";
 import { tw } from "~/utils";
 import { DeleteBooking } from "./delete-booking";
 import { Button } from "../shared";
