+++ app/components/booking/index.tsx
@@ -1,1 +1,1 @@
-export * from "./form";
+export * from "./page-content";
